requirements.txt
# FastAPI dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic
pydantic-settings==2.0.3
starlette==0.27.0
python-multipart==0.0.6
email-validator==2.0.0

# Database
sqlalchemy==2.0.22
alembic==1.12.0
psycopg2-binary

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
requests==2.31.0
qrcode[pil]==7.4.2
python-dateutil==2.8.2

# Testing
pytest==7.4.3
httpx==0.25.0